Split the CUDA cleanup handler to another file, add a param to control it.
